Split propolis-server config into separate crate
