+++ crates/propolis-server-config/Cargo.toml
@@ -1,0 +1,13 @@
+[package]
+name = "propolis-server-config"
+version = "0.0.0"
+license = "MPL-2.0"
+edition = "2018"
+
+# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+
+[dependencies]
+serde = "1.0"
+serde_derive = "1.0"
+toml = "0.5"
+thiserror = "1.0"
